+++ packages/regal/lib/src/theme/theme.dart
@@ -2,8 +2,7 @@ export 'brightness_extension.dart';
 export 'colors/color_system.dart';
 export 'colors/colors.dart';
 export 'colors/regal_colors.dart';
-export 'ledger_colors/ledger_colors.dart'
-    show LedgerColorX, NeutralColors, BaseColors;
+
 export 'text/text.dart';
 export 'theme_data.dart';
 export 'theme_extension.dart';