+++ packages/regal/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 3.8.0
+
+## 🚀 Updates:
+
+- Ledger DLS
+
+# 3.7.8
+
+## 🚀 Updates:
+
+- Exposed `IntactColors`
+
+# 3.7.7
+
+## 💅 Polish:
+
+- Update Legder Buttons and Regal Search Bar.
+
 # 3.7.6
 
 ## 💅 Polish: